playing with cards…
The old MMA site tries to use heading sizes to connote chunks /
sections of a page – i.e. stuff that is pertinent to a borrower. This
doesn’t seem to work very well, so I’m here experimenting with using
Materialize’s card styling to highlight specific sections of
information.